Azure DevOps integration
Setting up basic CI in Azure Pipelines, though
implementing it certainly posed a challenge...
not because it is hard, but hard to know all the
quirks. Ultimately I ended up creating 2 containers
which definitely slows down the build process,
however it renders the build predictable and
reproducible.

There are more enhancements on the way.

Bumping version as well, though no code changes.