Adding ability to publish host accessory
After the last refactor it was quite easy to actually
implement this... I expected more brainwrecking solutions
and a ton of bugs. But I encountered neither.

So at the end I toyed a little bit with #13 as well, and
I might have figured out a good way to do the integration test
without actually using HomeBridge. We'll see.

Anyway, this implements #3.