Rewriting almost all of homebridge-unraid
Honestly this became a lot bigger than I initially expected, but I have
no regrets. The code is still a bit messy, but it will be a good base
for future gardening. This will speed up development on multple upcoming
features.

The biggest change (from user perspective) is that containers and VMS
(along with the server controlling switch) will now be included in a
single accessory with multiple services. You have the option in the Home
app to present them separately, so you can get back to the old style
if you want.
Due to this reorg on the HomeKit level accessories are not recreated
upon restarts or updates, so you have the option to use shortcuts on
them, without the added fear of eventually breaking them.
It will still be a bit messy until automatic host control is introduced,
but overall I'm glad I got this out of the pipeline.

Existing config needs to be updated, though migration is provided.

Notable changes on the code level:
- Reducing the number (and size) of dependencies
- Removing HomeBridge and HAP as a necessary dependency (should result in
a leaner production setup)
- Introducing a bunch of utilities that makes future development easier
- A complete re-write :)