Implementing automatic host power management
When the server provides services that does not need to run 24/7
(e.g. VMs utilised only from time-to-time) it was a bother always
switching the host on and off when it was necessary. This feature
adds control so that underlying services can control the host
machine's power state as well - switch on when the service is
turned on, and switch off when it's not necessary anymore.

- Implements #12
- Fixes regression disregarding switchOffMechanism in config
- Fixes some inconsistency and unreasonable names in code