Bumping homebridge version, and with that a little refactor
Bumping the HB version we depend on means we're finally utilizing the
TypeScript API. Therefore I could simplify a bunch of things, while some
workarounds needed to stay in.

In the meantime I think there have been multiple things in the wrong
place, so we needed to shake it up a bit.